動画再生終了後のキャッシュ再開処理の再対応、それとキャッシュ中アイテムのキャンセル操作の再対応

+++ NicoPlayerHohoema/Services/Commands/Cache/DeleteCacheRequestCommand.cs
@@ -26,18 +26,20 @@ namespace NicoPlayerHohoema.Commands.Cache
             {
                 var video = Database.NicoVideoDb.Get(content.Id);
                 var cacheRequests = await VideoCacheManager.GetCacheRequest(content.Id);
-                var choiceItems = await DialogService.ShowMultiChoiceDialogAsync(
-                    $"削除するキャッシュ動画を選択する\n「{video.Title}」",
-                    cacheRequests,
-                    Enumerable.Empty<Models.Cache.NicoVideoCacheRequest>(),
-                    nameof(Models.Cache.NicoVideoCacheRequest.Quality)
-                    );
-
-                if (choiceItems?.Any() ?? false)
+                if (cacheRequests.Any())
                 {
-                    foreach (var deleteItem in choiceItems)
+                    var choiceItems = await DialogService.ShowMultiChoiceDialogAsync(
+                        $"削除するキャッシュ動画を選択する\n「{video.Title}」",
+                        cacheRequests,
+                        Enumerable.Empty<Models.Cache.NicoVideoCacheRequest>(),
+                        nameof(Models.Cache.NicoVideoCacheRequest.Quality)
+                        );
+                    if (choiceItems?.Any() ?? false)
                     {
-                        await VideoCacheManager.DeleteCachedVideo(content.Id, deleteItem.Quality);
+                        foreach (var deleteItem in choiceItems)
+                        {
+                            await VideoCacheManager.CancelCacheRequest(content.Id, deleteItem.Quality);
+                        }
                     }
                 }
             }
@@ -47,7 +49,12 @@ namespace NicoPlayerHohoema.Commands.Cache
         {
             if (parameter is Interfaces.IVideoContent content)
             {
-                return VideoCacheManager.CheckCached(content.Id);
+                // Note: CheckCacheRequested はキャッシュ済みしか同期実行で返せない
+                // これはVideoCacheManagerの設計が複雑で、
+                // 「キャッシュ状態ごとの動画」を非同期に列挙する形になってることが原因
+                // 本来は「動画ごとのキャッシュ状態」を非同期安全な形で同期的に取り出せる必要がある
+                //return VideoCacheManager.CheckCacheRequested(content.Id);
+                return true;
             }
             else
             {
